refactor: move utilities to separate package
