+++ eslint.config.js
@@ -5,7 +5,6 @@ import svelte from "eslint-plugin-svelte";
 import globals from "globals";
 import { fileURLToPath } from "node:url";
 import ts from "typescript-eslint";
-import svelteConfig from "./svelte.config.js";
 
 const gitignorePath = fileURLToPath(new URL("./.gitignore", import.meta.url));
 
@@ -30,7 +29,6 @@ export default ts.config(
                 projectService: true,
                 extraFileExtensions: [".svelte"],
                 parser: ts.parser,
-                svelteConfig,
             },
         },
     },
