+++ package.json
@@ -1,0 +1,16 @@
+{
+	"devDependencies": {
+		"@eslint/compat": "^1.2.9",
+		"@eslint/js": "^9.27.0",
+		"@tsconfig/svelte": "^5.0.4",
+		"eslint": "^9.27.0",
+		"eslint-config": "^0.3.0",
+		"eslint-config-prettier": "^10.1.5",
+		"eslint-plugin-svelte": "^3.9.0",
+		"globals": "^16.2.0",
+		"prettier": "^3.5.3",
+		"prettier-plugin-svelte": "^3.4.0",
+		"typescript": "^5.8.3",
+		"typescript-eslint": "^8.33.0"
+	}
+}
