+++ packages/utils/src/copy.ts
@@ -1,0 +1,10 @@
+export async function copy(text: string) {
+    try {
+        await navigator.clipboard.writeText(text);
+    } catch (err) {
+        if (err instanceof Error) {
+            console.error(err);
+        }
+        throw err;
+    }
+}
