+++ packages/utils/src/elementIndex.ts
