+++ userscripts/leetcode-toolkit/src/lib/jupyter.ts
@@ -5,7 +5,7 @@ import type {
     MultilineString,
 } from "@jupyterlab/nbformat";
 import { globalState } from "./state";
-import { downloadFile } from "./utils/download-file";
+import { downloadFile } from "@userscripts/utils/src/download-file";
 
 export function createNotebook({
     title,
