+++ userscripts/leetcode-toolkit/src/routes/problems/description/index.ts
@@ -1,5 +1,5 @@
 import { CONFIG } from "$lib/config";
-import { findElement } from "$lib/utils/elementFinder";
+import { findElement } from "@userscripts/utils/src/elementFinder";
 import { toKebabCase } from "remeda";
 import { mount } from "svelte";
 import DescriptionButttons from "./DescriptionButttons.svelte";
