+++ userscripts/leetcode-toolkit/tsconfig.json
@@ -1,10 +1,7 @@
 {
-    "extends": "@tsconfig/svelte/tsconfig.json",
+    "extends": "../../tsconfig.json",
     "compilerOptions": {
-        "target": "ESNext",
         "useDefineForClassFields": true,
-        "module": "ESNext",
-        "resolveJsonModule": true,
         "baseUrl": ".",
         "paths": {
             "$lib/*": ["src/lib/*"]
